requirements.txt
nats-python==0.8.0
asyncio-nats-client==0.10.0
asyncio-nats-streaming==0.4.0

